Reorientation/CardGen
"Sidebar" was deemed irrelevant to my objective, and generally not useful. Readding previous files except Motivation and Affirmation, both of which are replaced with CardGen, which uses code shared between the two files and generates cards for the right side of the page in one file. page.tsx also edited as necessary.
Not confirmed to be final product, waiting on edits to Motivation. Potential plans to include methods for objects (including Breathing and Analytics).